update text and move TextMessage to my_typing file
