+++ src/my_typing.py
@@ -86,3 +86,43 @@ class ChangesList(list):
                 if elem.type == __value.type and elem.second == __value.second and elem.weekday == __value.weekday:
                     return index
         super().index(__value, __start, __stop)
+
+
+# Порядок ввода: tuple(краткое имя, русское вариант) ......
+class TextMessage:
+    __instance = None
+
+    def __new__(cls, *args, **kwargs):
+        if cls.__instance is None:
+            cls.__instance = super().__new__(cls)
+
+        return cls.__instance
+
+    def __init__(self, *args):
+        self.__text = self.__args_to_dict(args)
+
+    @property
+    def text(self):
+        return self.__text
+
+    @text.setter
+    def text(self, value):
+        raise ValueError('You can not change this attribute')
+
+    @classmethod
+    def __args_to_dict(cls, arguments: tuple):
+        res = {'ru': {},
+               'en': {}}
+
+        for i in arguments:
+            for index, value in enumerate(i):
+                if index % 3 == 1:
+                    res['ru'][i[0]] = value
+                elif index % 3 == 2:
+                    res['en'][i[0]] = value
+                else:
+                    continue
+        return res
+
+    def __call__(self, short_name_text_mes: str, lang: str):
+        return self.__text[lang][short_name_text_mes]
