+++ src/tgbot/text.py
@@ -1,43 +1,4 @@
-# Порядок ввода: tuple(краткое имя, русское вариант) ......
-# Перевод на англ, автоматический
-class TextMessage:
-    __instance = None
-
-    def __new__(cls, *args, **kwargs):
-        if cls.__instance is None:
-            cls.__instance = super().__new__(cls)
-
-        return cls.__instance
-
-    def __init__(self, *args):
-        self.__text = self.__args_to_dict(args)
-
-    @property
-    def text(self):
-        return self.__text
-
-    @text.setter
-    def text(self, value):
-        raise ValueError('You can not change this attribute')
-
-    @classmethod
-    def __args_to_dict(cls, arguments: tuple):
-        res = {'ru': {},
-               'en': {}}
-
-        for i in arguments:
-            for index, value in enumerate(i):
-                if index % 3 == 1:
-                    res['ru'][i[0]] = value
-                elif index % 3 == 2:
-                    res['en'][i[0]] = value
-                else:
-                    continue
-        return res
-
-    def __call__(self, short_name_text_mes: str, lang: str):
-        return self.__text[lang][short_name_text_mes]
-
+from src.my_typing import TextMessage
 
 TEXT = TextMessage(('welcome', 'Добро пожаловать, ', 'Welcome, '),
                    ('hello', '✅ Пройди регистрацию в пару кликов', '✅ Register in a couple of clicks:'),
@@ -137,4 +98,13 @@ TEXT = TextMessage(('welcome', 'Добро пожаловать, ', 'Welcome, ')
                    ("administrator_dismissed", 'Ваша заявка была отклонена', 'Your application has been rejected'),
                    ("new_administrator_text", 'Ваша заявка была одобрена', 'Your application has been approved'),
                    ('admin_panel', 'Функции: ', 'Functions: '),
-                   ("change_schedule", "Изменить расписание", "Change schedule"))
+                   ("change_schedule", "Изменить расписание", "Change schedule"),
+                   ('reg_error', 'Кажется, что-то пошло не так :( Мы уже разбираемся с этим. Попробуйте '
+                                 'зарегистрироваться еще раз (с помощью команды /start) и если ошибка повторится, '
+                                 'то напишите, пожалуйста, об этом нам в /feedback', "It seems something went wrong "
+                                                                                     ":( We are already looking into "
+                                                                                     "this. Try registering again ("
+                                                                                     "using the /start command) and "
+                                                                                     "if the error persists, "
+                                                                                     "please write to us about it at "
+                                                                                     "/feedback"))
